+++ lib/src/main_widget.dart
@@ -4,8 +4,8 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
+import 'package:google_maps_widget/src/models/direction.dart';
 import 'package:google_maps_widget/src/models/marker_icon_info.dart';
-import 'package:google_maps_widget/src/services/maps_service.dart';
 import 'package:google_maps_widget/src/utils/constants.dart';
 
 /// A [GoogleMapsWidget] which can be used to make polylines(route)
@@ -20,33 +20,23 @@ class GoogleMapsWidget extends StatefulWidget {
     this.totalDistanceCallback,
     this.totalTimeCallback,
     this.onMapCreated,
-    this.sourceMarkerIconInfo,
-    this.destinationMarkerIconInfo,
-    this.driverMarkerIconInfo,
-    this.onTapSourceMarker,
-    this.onTapDestinationMarker,
-    this.onTapDriverMarker,
-    this.onTapSourceInfoWindow,
-    this.onTapDestinationInfoWindow,
-    this.onTapDriverInfoWindow,
+    this.sourceMarkerIconInfo = const MarkerIconInfo(),
+    this.destinationMarkerIconInfo = const MarkerIconInfo(),
+    this.driverMarkerIconInfo = const MarkerIconInfo(),
     this.driverCoordinatesStream,
     this.defaultCameraLocation,
+    this.onPolylineUpdate,
     this.markers = const <Marker>{},
     this.polylines = const <Polyline>{},
     this.showPolyline = true,
-    this.showSourceMarker = true,
-    this.showDestinationMarker = true,
-    this.showDriverMarker = true,
     this.defaultCameraZoom = Constants.kDefaultCameraZoom,
-    this.sourceName = Constants.kDefaultSourceName,
-    this.destinationName = Constants.kDefaultDestinationName,
-    this.driverName = Constants.kDefaultDriverName,
     this.routeColor = Constants.kRouteColor,
     this.routeWidth = Constants.kRouteWidth,
     this.updatePolylinesOnDriverLocUpdate = true,
 
     // other google maps params
     this.gestureRecognizers = const <Factory<OneSequenceGestureRecognizer>>{},
+    this.layoutDirection,
     this.compassEnabled = true,
     this.mapToolbarEnabled = true,
     this.cameraTargetBounds = CameraTargetBounds.unbounded,
@@ -82,24 +72,6 @@ class GoogleMapsWidget extends StatefulWidget {
 
   /// The destination [LatLng].
   final LatLng destinationLatLng;
-
-  /// Called every time source [Marker]'s [InfoWindow] is tapped.
-  final void Function(LatLng)? onTapSourceInfoWindow;
-
-  /// Called every time destination [Marker]'s [InfoWindow] is tapped.
-  final void Function(LatLng)? onTapDestinationInfoWindow;
-
-  /// Called every time driver [Marker]'s [InfoWindow] is tapped.
-  final void Function(LatLng)? onTapDriverInfoWindow;
-
-  /// Called every time source [Marker] is tapped.
-  final void Function(LatLng)? onTapSourceMarker;
-
-  /// Called every time destination [Marker] is tapped.
-  final void Function(LatLng)? onTapDestinationMarker;
-
-  /// Called every time driver [Marker] is tapped.
-  final void Function(LatLng)? onTapDriverMarker;
 
   /// If true, Updates the polylines everytime a new event is pushed to
   /// the driver stream, i.e. the driver location changes..
@@ -137,21 +109,6 @@ class GoogleMapsWidget extends StatefulWidget {
   /// Defaults to [Constants.kDefaultCameraZoom].
   final double defaultCameraZoom;
 
-  /// Displays source [Marker]'s [InfoWindow] displaying [sourceName]
-  /// when tapped on [sourceMarkerIconInfo].
-  /// Defaults to [Constants.kDefaultSourceName].
-  final String sourceName;
-
-  /// Displays destination [Marker]'s [InfoWindow] displaying [destinationName]
-  /// when tapped on [destinationMarkerIconInfo].
-  /// Defaults to [Constants.kDefaultDestinationName].
-  final String destinationName;
-
-  /// Displays driver's [Marker]'s [InfoWindow] displaying [driverName]
-  /// when tapped on [driverMarkerIconInfo].
-  /// Defaults to [Constants.kDefaultDriverName].
-  final String driverName;
-
   /// Color of the route made between [sourceLatLng] and [destinationLatLng].
   /// Defaults to [Constants.kRouteColor].
   final Color routeColor;
@@ -160,33 +117,21 @@ class GoogleMapsWidget extends StatefulWidget {
   /// Defaults to [Constants.kRouteWidth].
   final int routeWidth;
 
+  /// Called when the polyline is updated i.e. route is updated.
+  final void Function(Polyline)? onPolylineUpdate;
+
   /// The marker which is rendered on the location [sourceLatLng].
-  final MarkerIconInfo? sourceMarkerIconInfo;
+  final MarkerIconInfo sourceMarkerIconInfo;
 
   /// The marker which is rendered on the location [destinationLatLng].
-  final MarkerIconInfo? destinationMarkerIconInfo;
+  final MarkerIconInfo destinationMarkerIconInfo;
 
   /// The marker which is rendered on the driver's current location
   /// provided by [driverCoordinatesStream].
   ///
   /// See also:
   ///   * [driverCoordinatesStream] parameter.
-  final MarkerIconInfo? driverMarkerIconInfo;
-
-  /// Whether to show the source marker at [sourceLatLng].
-  ///
-  /// Defaults to true.
-  final bool showSourceMarker;
-
-  /// Whether to show the destination marker at [destinationLatLng].
-  ///
-  /// Defaults to true.
-  final bool showDestinationMarker;
-
-  /// Whether to show the driver marker.
-  ///
-  /// Defaults to true.
-  final bool showDriverMarker;
+  final MarkerIconInfo driverMarkerIconInfo;
 
   /// Whether to show the generated polyline from [sourceLatLng]
   /// to [destinationLatLng].
@@ -273,6 +218,12 @@ class GoogleMapsWidget extends StatefulWidget {
   /// Enables or disables showing 3D buildings where available
   final bool buildingsEnabled;
 
+  /// The layout direction to use for the embedded view.
+  ///
+  /// If this is null, the ambient [Directionality] is used instead. If there is
+  /// no ambient [Directionality], [TextDirection.ltr] is used.
+  final TextDirection? layoutDirection;
+
   /// True if the map should show a compass when rotated.
   final bool compassEnabled;
 
@@ -350,51 +301,201 @@ class GoogleMapsWidget extends StatefulWidget {
   final CameraTargetBounds cameraTargetBounds;
 
   @override
-  // ignore: library_private_types_in_public_api
-  _GoogleMapsWidgetState createState() => _GoogleMapsWidgetState();
+  GoogleMapsWidgetState createState() => GoogleMapsWidgetState();
 }
 
-class _GoogleMapsWidgetState extends State<GoogleMapsWidget> {
-  final _mapsService = MapsService();
+class GoogleMapsWidgetState extends State<GoogleMapsWidget> {
+  /// Markers to be placed on the map.
+  final _markersMap = <MarkerId, Marker>{};
+
+  /// Polylines to be placed on the map.
+  final _polylines = <Polyline>{};
+
+  final _mapsControllerCompleter = Completer<GoogleMapController>();
+
+  Future<GoogleMapController> getGoogleMapsController() =>
+      _mapsControllerCompleter.future;
+
+  Future<void> setSourceLatLng(LatLng sourceLatLng) async {
+    if (_sourceLatLng == sourceLatLng) return;
+
+    _sourceLatLng = sourceLatLng;
+    await _buildPolyLines();
+    _setSourceDestinationMarkers();
+    setState(() {});
+  }
+
+  Future<void> setDestinationLatLng(LatLng destinationLatLng) async {
+    if (_destinationLatLng == destinationLatLng) return;
+
+    _destinationLatLng = destinationLatLng;
+    await _buildPolyLines();
+    _setSourceDestinationMarkers();
+    setState(() {});
+  }
+
+  StreamSubscription<LatLng>? _driverCoordinatesStreamSubscription;
+
+  late LatLng _sourceLatLng;
+  late LatLng _destinationLatLng;
+
+  /// setting source and destination markers
+  void _setSourceDestinationMarkers() async {
+    _markersMap.remove(MarkerIconInfo.sourceMarkerId);
+
+    final sourceMarker = widget.sourceMarkerIconInfo;
+    if (sourceMarker.isVisible) {
+      _markersMap[MarkerIconInfo.sourceMarkerId] = Marker(
+        markerId: MarkerIconInfo.sourceMarkerId,
+        position: _sourceLatLng,
+        anchor: sourceMarker.anchor,
+        rotation: sourceMarker.rotation,
+        icon: await sourceMarker.bitmapDescriptor,
+        onTap: sourceMarker.onTapMarker == null
+            ? null
+            : () => sourceMarker.onTapMarker!(_sourceLatLng),
+        infoWindow: InfoWindow(
+          onTap: sourceMarker.onTapInfoWindow == null
+              ? null
+              : () => sourceMarker.onTapInfoWindow!(_sourceLatLng),
+          title: sourceMarker.infoWindowTitle,
+        ),
+      );
+    }
+
+    _markersMap.remove(MarkerIconInfo.destinationMarkerId);
+
+    final destinationMarker = widget.destinationMarkerIconInfo;
+    if (destinationMarker.isVisible) {
+      _markersMap[MarkerIconInfo.destinationMarkerId] = Marker(
+        markerId: MarkerIconInfo.destinationMarkerId,
+        position: _destinationLatLng,
+        anchor: destinationMarker.anchor,
+        rotation: destinationMarker.rotation,
+        icon: await destinationMarker.bitmapDescriptor,
+        onTap: destinationMarker.onTapMarker == null
+            ? null
+            : () => destinationMarker.onTapMarker!(_destinationLatLng),
+        infoWindow: InfoWindow(
+          onTap: destinationMarker.onTapInfoWindow == null
+              ? null
+              : () => destinationMarker.onTapInfoWindow!(_destinationLatLng),
+          title: destinationMarker.infoWindowTitle,
+        ),
+      );
+    }
+
+    setState(() {});
+  }
+
+  /// Build polylines from [_sourceLatLng] to [_destinationLatLng].
+  Future<void> _buildPolyLines({LatLng? driverLoc}) async {
+    if (!widget.showPolyline) return;
+
+    final result = await Direction.getDirections(
+      googleMapsApiKey: widget.apiKey,
+      origin: driverLoc ?? _sourceLatLng,
+      destination: _destinationLatLng,
+    );
+
+    final polylineCoordinates = <LatLng>[];
+
+    if (result != null && result.polylinePoints.isNotEmpty) {
+      polylineCoordinates.addAll(result.polylinePoints);
+    }
+
+    final polyline = Polyline(
+      polylineId: const PolylineId('default-polyline'),
+      color: widget.routeColor,
+      width: widget.routeWidth,
+      points: polylineCoordinates,
+    );
+
+    widget.onPolylineUpdate?.call(polyline);
+
+    if (driverLoc != null) _polylines.clear();
+    _polylines.add(polyline);
+
+    // setting map such as both source and
+    // destinations markers can be seen
+    if (result != null) {
+      final controller = await getGoogleMapsController();
+
+      controller.animateCamera(
+        CameraUpdate.newLatLngBounds(result.bounds, 32),
+      );
+
+      widget.totalTimeCallback?.call(result.totalDuration);
+      widget.totalDistanceCallback?.call(result.totalDistance);
+    }
+
+    setState(() {});
+  }
+
+  /// This function uses [GoogleMapsWidget.driverCoordinatesStream] which
+  /// is a [Stream] of [LatLng]-[coordinates] and renders
+  /// [GoogleMapsWidget.driverMarkerIconInfo] marker to show
+  /// driver's location in realtime.
+  Future<void> _listenToDriverCoordinates() async {
+    final driverStream = widget.driverCoordinatesStream;
+    if (driverStream == null) return;
+
+    final driverMarkerBitmapDescriptor =
+        await widget.driverMarkerIconInfo.bitmapDescriptor;
+
+    _driverCoordinatesStreamSubscription = driverStream.listen((coordinate) {
+      if (widget.updatePolylinesOnDriverLocUpdate) {
+        _buildPolyLines(driverLoc: coordinate);
+      }
+
+      if (!widget.driverMarkerIconInfo.isVisible) return;
+
+      _markersMap.remove(MarkerIconInfo.driverMarkerId);
+
+      final driverMarker = widget.driverMarkerIconInfo;
+      _markersMap[MarkerIconInfo.driverMarkerId] = Marker(
+        markerId: MarkerIconInfo.driverMarkerId,
+        position: coordinate,
+        anchor: driverMarker.anchor,
+        rotation: driverMarker.rotation,
+        icon: driverMarkerBitmapDescriptor,
+        onTap: driverMarker.onTapMarker == null
+            ? null
+            : () => driverMarker.onTapMarker!(coordinate),
+        infoWindow: InfoWindow(
+          onTap: driverMarker.onTapInfoWindow == null
+              ? null
+              : () => driverMarker.onTapInfoWindow!(coordinate),
+          title: driverMarker.infoWindowTitle,
+        ),
+      );
+
+      setState(() {});
+    });
+  }
 
   @override
   void initState() {
-    _mapsService.initialize(
-      setState: setState,
-      apiKey: widget.apiKey,
-      sourceLatLng: widget.sourceLatLng,
-      destinationLatLng: widget.destinationLatLng,
-      onTapSourceMarker: widget.onTapSourceMarker,
-      onTapDestinationMarker: widget.onTapDestinationMarker,
-      onTapDriverMarker: widget.onTapDriverMarker,
-      onTapSourceInfoWindow: widget.onTapSourceInfoWindow,
-      onTapDestinationInfoWindow: widget.onTapDestinationInfoWindow,
-      onTapDriverInfoWindow: widget.onTapDriverInfoWindow,
-      driverCoordinatesStream: widget.driverCoordinatesStream,
-      updatePolylinesOnDriverLocUpdate: widget.updatePolylinesOnDriverLocUpdate,
-      sourceName: widget.sourceName,
-      destinationName: widget.destinationName,
-      driverName: widget.driverName,
-      routeColor: widget.routeColor,
-      routeWidth: widget.routeWidth,
-      defaultCameraLocation: widget.defaultCameraLocation,
-      defaultCameraZoom: widget.defaultCameraZoom,
-      sourceMarkerIconInfo: widget.sourceMarkerIconInfo,
-      destinationMarkerIconInfo: widget.destinationMarkerIconInfo,
-      driverMarkerIconInfo: widget.driverMarkerIconInfo,
-      totalTimeCallback: widget.totalTimeCallback,
-      totalDistanceCallback: widget.totalDistanceCallback,
-      showSourceMarker: widget.showSourceMarker,
-      showDestinationMarker: widget.showDestinationMarker,
-      showDriverMarker: widget.showDriverMarker,
-      showPolyline: widget.showPolyline,
-    );
+    _sourceLatLng = widget.sourceLatLng;
+    _destinationLatLng = widget.destinationLatLng;
+
+    _setSourceDestinationMarkers();
+    _buildPolyLines();
+    _listenToDriverCoordinates();
+
     super.initState();
   }
 
   @override
   void dispose() {
-    _mapsService.clear();
+    _driverCoordinatesStreamSubscription?.cancel();
+
+    if (_mapsControllerCompleter.isCompleted) {
+      _mapsControllerCompleter.future.then(
+        (controller) => controller.dispose(),
+      );
+    }
+
     super.dispose();
   }
 
@@ -402,13 +503,13 @@ class _GoogleMapsWidgetState extends State<GoogleMapsWidget> {
   Widget build(BuildContext context) {
     return GoogleMap(
       initialCameraPosition: CameraPosition(
-        target: _mapsService.defaultCameraLocation,
-        zoom: _mapsService.defaultCameraZoom,
+        target: widget.defaultCameraLocation ?? _sourceLatLng,
+        zoom: widget.defaultCameraZoom,
       ),
-      markers: {..._mapsService.markers, ...widget.markers},
-      polylines: {..._mapsService.polylines, ...widget.polylines},
+      markers: {..._markersMap.values, ...widget.markers},
+      polylines: {..._polylines, ...widget.polylines},
       onMapCreated: (controller) {
-        _mapsService.setController(controller);
+        _mapsControllerCompleter.complete(controller);
         if (widget.onMapCreated != null) {
           return widget.onMapCreated!(controller);
         }
@@ -418,6 +519,7 @@ class _GoogleMapsWidgetState extends State<GoogleMapsWidget> {
       /////////////////////////////////////////////////
       gestureRecognizers: widget.gestureRecognizers,
       compassEnabled: widget.compassEnabled,
+      layoutDirection: widget.layoutDirection,
       mapToolbarEnabled: widget.mapToolbarEnabled,
       cameraTargetBounds: widget.cameraTargetBounds,
       mapType: widget.mapType,